settings done

+++ App/Models/IncomeDB.php
@@ -109,8 +109,27 @@ class IncomeDB extends \Core\Model
 
     public function deleteIncomesCategory()
     {
-        $sql = 'DELETE FROM incomes_category_assigned_to_users
+        if ($this->deleteIncomesCategoryCD()) {
+            $sql = 'DELETE FROM incomes_category_assigned_to_users
         WHERE incomes_category_assigned_to_users.name = :name AND user_id = :user_id';
+
+            $db = static::getDB();
+            $stmt = $db->prepare($sql);
+
+            $stmt->bindValue(':name', $this->incomeCategory, PDO::PARAM_STR);
+            $stmt->bindValue(':user_id', $_SESSION['user_id'], PDO::PARAM_INT);
+
+            return $stmt->execute();
+        }
+        return false;
+    }
+
+    public function deleteIncomesCategoryCD()
+    {
+        $sql = 'DELETE incomes FROM incomes
+        INNER JOIN incomes_category_assigned_to_users ON
+        incomes_category_assigned_to_users.id = incomes.income_category_assigned_to_user_id
+        WHERE incomes_category_assigned_to_users.name = :name AND incomes.user_id = :user_id';
 
         $db = static::getDB();
         $stmt = $db->prepare($sql);
@@ -125,6 +144,10 @@ class IncomeDB extends \Core\Model
     {
         if (NULL != $this->findTheSameNameOfCategory()) {
             $this->errors[] = 'Taka kategoria już istnieje - wpisz inną nazwę';
+        }
+
+        if (NULL == $this->newIncomeCategory) {
+            $this->errors[] = 'Nie można dodać pustej kategorii - wpisz nazwę.';
         }
     }
